requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
braceexpand==0.1.7
charset-normalizer==2.1.0
click==8.1.3
discord.py @ git+https://github.com/Rapptz/discord.py@5538cd501c4055c12c8d78534a8d8c648fef125c
dnspython==2.2.1
frozenlist==1.3.1
idna==3.3
import-expression==1.1.4
importlib-metadata==4.12.0
jishaku==2.5.1
line-profiler==3.5.1
multidict==6.0.2
Pillow==9.2.0
pymongo==4.2.0
six==1.16.0
typing_extensions==4.3.0
yarl==1.8.1
zipp==3.8.1
